DB-4890: Check if cache_discovery table exists.

+++ profiles/pantheon_decoupled_profile/pantheon_decoupled_profile.profile
@@ -33,7 +33,9 @@ function pantheon_decoupled_install_demo_content(array &$install_state) {
 
   // @TODO: This manual cache clearing should not be necessary,
   // if https://www.drupal.org/project/drupal/issues/3076544 is fixed.
+  if (\Drupal::database()->schema()->tableExists('cache_discovery')) {
     \Drupal::database()->truncate('cache_discovery')->execute();
+  }
 
   \Drupal::service('module_installer')->install(['pantheon_decoupled_example']);
 }
